TYPE: merge

KEYWORDS: WRFPLUS v3.8.1, WRFPLUS-Chem v3.6.1

WARNING: This merge commit has not been compiled or tested yet.  There are probably bugs.

DESCRIPTION:
This merge brings WRFPLUS-Chem almost up-to-date with WRFPLUS.  The exception is that the following NL model files were modified in the intervening period to imrpove physics:
	phys/module_sf_pxsfclay.F
	phys/module_sf_pxlsm.F
	phys/module_bl_acm.F
The old verions of these files were kept, along with some modification to phys/module_surface_driver.F
and phys/module_pbl_driver.F in order to maintain agreement between the NL, TL, and AD
models (i.e. module_bl_acm_tl.F, module_bl_acm_ad.F, etc.). These modifications
are annotated for future archaeologists or eager grad students. Further work is necessary to
update the TL and AD code for consistency with the improved NL versions.

Conflicts were manually resolved for:
	Registry/Registry.EM_COMMON
	Registry/registry.wrfplus
	arch/configure_new.defaults
	chem/Makefile
	chem/aerosol_driver.F
	chem/chem_driver.F
	chem/chemics_init.F
	chem/depend.chem
	chem/emissions_driver.F
	chem/mechanism_driver.F
	chem/module_chem_plumerise_scalar.F
	chem/module_emissions_anthropogenics.F
	chem/module_input_chem_data.F
	chem/module_plumerise1.F
	dyn_em/module_first_rk_step_part1.F
	dyn_em/module_first_rk_step_part1_ad.F
	dyn_em/module_first_rk_step_part1_tl.F
	dyn_em/solve_em_ad.F
	dyn_em/solve_em_tl.F
	frame/module_integrate.F
	main/module_wrf_top.F
	phys/module_bl_acm.F
	phys/module_pbl_driver_ad.F
	phys/module_pbl_driver_tl.F
	phys/module_radiation_driver.F
	phys/module_sf_pxlsm.F
	phys/module_sf_pxsfclay.F
	share/mediation_integrate.F
	share/module_adtl_grid_utilities.F

+++ var/scripts/gen_be/gen_be_stage0_wrf.ksh
@@ -29,15 +29,24 @@ export START_DATE_STAGE0=$($DAAT $START_DATE -$FCST_RANGE1)
 export END_DATE_STAGE0=$($DAAT $END_DATE   -$FCST_RANGE1)
 export DATE=$START_DATE_STAGE0
 
+if [[ $START_DATE_STAGE0 > $END_DATE_STAGE0 ]]; then
+   echo "Start date $START_DATE_STAGE0 must be before $END_DATE_STAGE0!"
+   exit -1
+fi
+
+export CV_OPTIONS=$NL_CV_OPTIONS #Set CV_OPTIONS (either specified by wrapper or defaults)
+
 while [[ $DATE -le $END_DATE_STAGE0 ]]; do
    export TMP_DIR=${WORK_DIR}/${DATE}
    rm -rf ${TMP_DIR} 2>/dev/null
    mkdir ${TMP_DIR}  2>/dev/null
    cd ${TMP_DIR}
 
-   for SV in psi chi t rh ps; do
-      if [[ ! -d $SV ]]; then mkdir $SV; fi
-   done
+   if [[ $NL_CV_OPTIONS == 7 ]]; then
+      for SV in u v t rh ps; do mkdir -p $SV; done
+   else
+      for SV in psi chi t rh ps; do mkdir -p $SV; done
+   fi
 
    #  Create file dates:
    export FCST_TIME=$($DAAT $DATE $FCST_RANGE1)
@@ -47,7 +56,11 @@ while [[ $DATE -le $END_DATE_STAGE0 ]]; do
    export MM=$(echo $FCST_TIME | cut -c5-6)
    export DD=$(echo $FCST_TIME | cut -c7-8)
    export HH=$(echo $FCST_TIME | cut -c9-10)
-   export FILE_DATE=${YYYY}-${MM}-${DD}_${HH}:00:00
+      if $NOCOLONS; then
+         export FILE_DATE=${YYYY}-${MM}-${DD}_${HH}_00_00
+      else
+         export FILE_DATE=${YYYY}-${MM}-${DD}_${HH}:00:00
+      fi
    export FILE=${FC_DIR}/${DATE}/wrfout_d${DOMAIN}_${FILE_DATE}
    export FILE1=wrfout_d${DOMAIN}_${FILE_DATE}
    export NEXT_DATE=$($DAAT $DATE $INTERVAL)
@@ -68,7 +81,7 @@ while [[ $DATE -le $END_DATE_STAGE0 ]]; do
    fi
 
    ln -fs ${BUILD_DIR}/gen_be_stage0_wrf.exe .
-   ./gen_be_stage0_wrf.exe ${BE_METHOD} ${FCST_TIME} $NE $FILE1 > gen_be_stage0_wrf.${FCST_TIME}.log 2>&1
+   ./gen_be_stage0_wrf.exe ${BE_METHOD} ${FCST_TIME} $NE $FILE1 ${CV_OPTIONS} > gen_be_stage0_wrf.${FCST_TIME}.log 2>&1
 
    #  Tidy:
    mv pert.${FCST_TIME}* ${STAGE0_DIR}
